fix: move tests to another directory to prevent mocha discovery and remove abstract cursor generic constraint
