+++ .gitignore
@@ -56,7 +56,7 @@ lib/
 *.tgz
 *.d.ts
 # type definition tests
-!test/unit/types
+!test/types
 
 .vscode
 output
