+++ test/types/mongodb.test-d.ts
@@ -1,10 +1,10 @@
 import { expectDeprecated, expectType, expectNotType, expectError, expectAssignable } from 'tsd';
 
-import { Collection } from '../../.../../../src/collection';
-import { AggregationCursor } from '../../../src/cursor/aggregation_cursor';
-import { Db } from '../../../src/db';
-import { MongoError } from '../../../src/error';
-import { MongoClient } from '../../../src/mongo_client';
+import { Collection } from '../../src//collection';
+import { AggregationCursor } from '../../src//cursor/aggregation_cursor';
+import { Db } from '../../src//db';
+import { MongoError } from '../../src//error';
+import { MongoClient } from '../../src//mongo_client';
 
 expectDeprecated(Collection.prototype.insert);
 expectDeprecated(Collection.prototype.update);
